requirements.txt
numpy==1.25.2
opencv_python==4.8.0.76
Pillow==10.0.0
Pillow==10.0.1
PyAutoGUI==0.9.54
PyDirectInput==1.0.4
PyGetWindow==0.0.9
pynput==1.7.6
PyScreeze==0.1.29
pytesseract==0.3.10
python-dotenv==1.0.0
